LayoutTests/inspector/debugger/tail-recursion-expected.txt

Testing that we keep around tail deleted frames in the inspector.

Starting Test


------------------------------------
Hit breakpoint at line: 3, column: 8
numBreakpointHits: 1
Going to look at the top 50 frames.
------------------------------------
Looking at frame number: 0
    variable 'i': {"_type":"number","_description":"0","_hasChildren":false,"_value":0}
Looking at frame number: 1
    variable 'i': {"_type":"number","_description":"1","_hasChildren":false,"_value":1}
Looking at frame number: 2
    variable 'i': {"_type":"number","_description":"2","_hasChildren":false,"_value":2}
Looking at frame number: 3
    variable 'i': {"_type":"number","_description":"3","_hasChildren":false,"_value":3}
Looking at frame number: 4
    variable 'i': {"_type":"number","_description":"4","_hasChildren":false,"_value":4}
Looking at frame number: 5
    variable 'i': {"_type":"number","_description":"5","_hasChildren":false,"_value":5}
Looking at frame number: 6
    variable 'i': {"_type":"number","_description":"6","_hasChildren":false,"_value":6}
Looking at frame number: 7
    variable 'i': {"_type":"number","_description":"7","_hasChildren":false,"_value":7}
Looking at frame number: 8
    variable 'i': {"_type":"number","_description":"8","_hasChildren":false,"_value":8}
Looking at frame number: 9
    variable 'i': {"_type":"number","_description":"9","_hasChildren":false,"_value":9}
Looking at frame number: 10
    variable 'i': {"_type":"number","_description":"10","_hasChildren":false,"_value":10}
Looking at frame number: 11
    variable 'i': {"_type":"number","_description":"11","_hasChildren":false,"_value":11}
Looking at frame number: 12
    variable 'i': {"_type":"number","_description":"12","_hasChildren":false,"_value":12}
Looking at frame number: 13
    variable 'i': {"_type":"number","_description":"13","_hasChildren":false,"_value":13}
Looking at frame number: 14
    variable 'i': {"_type":"number","_description":"14","_hasChildren":false,"_value":14}
Looking at frame number: 15
    variable 'i': {"_type":"number","_description":"15","_hasChildren":false,"_value":15}
Looking at frame number: 16
    variable 'i': {"_type":"number","_description":"16","_hasChildren":false,"_value":16}
Looking at frame number: 17
    variable 'i': {"_type":"number","_description":"17","_hasChildren":false,"_value":17}
Looking at frame number: 18
    variable 'i': {"_type":"number","_description":"18","_hasChildren":false,"_value":18}
Looking at frame number: 19
    variable 'i': {"_type":"number","_description":"19","_hasChildren":false,"_value":19}
Looking at frame number: 20
    variable 'i': {"_type":"number","_description":"20","_hasChildren":false,"_value":20}
Looking at frame number: 21
    variable 'i': {"_type":"number","_description":"21","_hasChildren":false,"_value":21}
Looking at frame number: 22
    variable 'i': {"_type":"number","_description":"22","_hasChildren":false,"_value":22}
Looking at frame number: 23
    variable 'i': {"_type":"number","_description":"23","_hasChildren":false,"_value":23}
Looking at frame number: 24
    variable 'i': {"_type":"number","_description":"24","_hasChildren":false,"_value":24}
Looking at frame number: 25
    variable 'i': {"_type":"number","_description":"25","_hasChildren":false,"_value":25}
Looking at frame number: 26
    variable 'i': {"_type":"number","_description":"26","_hasChildren":false,"_value":26}
Looking at frame number: 27
    variable 'i': {"_type":"number","_description":"27","_hasChildren":false,"_value":27}
Looking at frame number: 28
    variable 'i': {"_type":"number","_description":"28","_hasChildren":false,"_value":28}
Looking at frame number: 29
    variable 'i': {"_type":"number","_description":"29","_hasChildren":false,"_value":29}
Looking at frame number: 30
    variable 'i': {"_type":"number","_description":"30","_hasChildren":false,"_value":30}
Looking at frame number: 31
    variable 'i': {"_type":"number","_description":"31","_hasChildren":false,"_value":31}
Looking at frame number: 32
    variable 'i': {"_type":"number","_description":"32","_hasChildren":false,"_value":32}
Looking at frame number: 33
    variable 'i': {"_type":"number","_description":"33","_hasChildren":false,"_value":33}
Looking at frame number: 34
    variable 'i': {"_type":"number","_description":"34","_hasChildren":false,"_value":34}
Looking at frame number: 35
    variable 'i': {"_type":"number","_description":"35","_hasChildren":false,"_value":35}
Looking at frame number: 36
    variable 'i': {"_type":"number","_description":"36","_hasChildren":false,"_value":36}
Looking at frame number: 37
    variable 'i': {"_type":"number","_description":"37","_hasChildren":false,"_value":37}
Looking at frame number: 38
    variable 'i': {"_type":"number","_description":"38","_hasChildren":false,"_value":38}
Looking at frame number: 39
    variable 'i': {"_type":"number","_description":"39","_hasChildren":false,"_value":39}
Looking at frame number: 40
    variable 'i': {"_type":"number","_description":"40","_hasChildren":false,"_value":40}
Looking at frame number: 41
    variable 'i': {"_type":"number","_description":"41","_hasChildren":false,"_value":41}
Looking at frame number: 42
    variable 'i': {"_type":"number","_description":"42","_hasChildren":false,"_value":42}
Looking at frame number: 43
    variable 'i': {"_type":"number","_description":"43","_hasChildren":false,"_value":43}
Looking at frame number: 44
    variable 'i': {"_type":"number","_description":"44","_hasChildren":false,"_value":44}
Looking at frame number: 45
    variable 'i': {"_type":"number","_description":"45","_hasChildren":false,"_value":45}
Looking at frame number: 46
    variable 'i': {"_type":"number","_description":"46","_hasChildren":false,"_value":46}
Looking at frame number: 47
    variable 'i': {"_type":"number","_description":"47","_hasChildren":false,"_value":47}
Looking at frame number: 48
    variable 'i': {"_type":"number","_description":"48","_hasChildren":false,"_value":48}
Looking at frame number: 49
    variable 'i': {"_type":"number","_description":"49","_hasChildren":false,"_value":49}


------------------------------------
Hit breakpoint at line: 4, column: 8
numBreakpointHits: 2
Going to look at the top 50 frames.
------------------------------------
Looking at frame number: 0
    variable 'i': {"_type":"number","_description":"1","_hasChildren":false,"_value":1}
Looking at frame number: 1
    variable 'i': {"_type":"number","_description":"1","_hasChildren":false,"_value":1}
Looking at frame number: 2
    variable 'i': {"_type":"number","_description":"2","_hasChildren":false,"_value":2}
Looking at frame number: 3
    variable 'i': {"_type":"number","_description":"3","_hasChildren":false,"_value":3}
Looking at frame number: 4
    variable 'i': {"_type":"number","_description":"4","_hasChildren":false,"_value":4}
Looking at frame number: 5
    variable 'i': {"_type":"number","_description":"5","_hasChildren":false,"_value":5}
Looking at frame number: 6
    variable 'i': {"_type":"number","_description":"6","_hasChildren":false,"_value":6}
Looking at frame number: 7
    variable 'i': {"_type":"number","_description":"7","_hasChildren":false,"_value":7}
Looking at frame number: 8
    variable 'i': {"_type":"number","_description":"8","_hasChildren":false,"_value":8}
Looking at frame number: 9
    variable 'i': {"_type":"number","_description":"9","_hasChildren":false,"_value":9}
Looking at frame number: 10
    variable 'i': {"_type":"number","_description":"10","_hasChildren":false,"_value":10}
Looking at frame number: 11
    variable 'i': {"_type":"number","_description":"11","_hasChildren":false,"_value":11}
Looking at frame number: 12
    variable 'i': {"_type":"number","_description":"12","_hasChildren":false,"_value":12}
Looking at frame number: 13
    variable 'i': {"_type":"number","_description":"13","_hasChildren":false,"_value":13}
Looking at frame number: 14
    variable 'i': {"_type":"number","_description":"14","_hasChildren":false,"_value":14}
Looking at frame number: 15
    variable 'i': {"_type":"number","_description":"15","_hasChildren":false,"_value":15}
Looking at frame number: 16
    variable 'i': {"_type":"number","_description":"16","_hasChildren":false,"_value":16}
Looking at frame number: 17
    variable 'i': {"_type":"number","_description":"17","_hasChildren":false,"_value":17}
Looking at frame number: 18
    variable 'i': {"_type":"number","_description":"18","_hasChildren":false,"_value":18}
Looking at frame number: 19
    variable 'i': {"_type":"number","_description":"19","_hasChildren":false,"_value":19}
Looking at frame number: 20
    variable 'i': {"_type":"number","_description":"20","_hasChildren":false,"_value":20}
Looking at frame number: 21
    variable 'i': {"_type":"number","_description":"21","_hasChildren":false,"_value":21}
Looking at frame number: 22
    variable 'i': {"_type":"number","_description":"22","_hasChildren":false,"_value":22}
Looking at frame number: 23
    variable 'i': {"_type":"number","_description":"23","_hasChildren":false,"_value":23}
Looking at frame number: 24
    variable 'i': {"_type":"number","_description":"24","_hasChildren":false,"_value":24}
Looking at frame number: 25
    variable 'i': {"_type":"number","_description":"25","_hasChildren":false,"_value":25}
Looking at frame number: 26
    variable 'i': {"_type":"number","_description":"26","_hasChildren":false,"_value":26}
Looking at frame number: 27
    variable 'i': {"_type":"number","_description":"27","_hasChildren":false,"_value":27}
Looking at frame number: 28
    variable 'i': {"_type":"number","_description":"28","_hasChildren":false,"_value":28}
Looking at frame number: 29
    variable 'i': {"_type":"number","_description":"29","_hasChildren":false,"_value":29}
Looking at frame number: 30
    variable 'i': {"_type":"number","_description":"30","_hasChildren":false,"_value":30}
Looking at frame number: 31
    variable 'i': {"_type":"number","_description":"31","_hasChildren":false,"_value":31}
Looking at frame number: 32
    variable 'i': {"_type":"number","_description":"32","_hasChildren":false,"_value":32}
Looking at frame number: 33
    variable 'i': {"_type":"number","_description":"33","_hasChildren":false,"_value":33}
Looking at frame number: 34
    variable 'i': {"_type":"number","_description":"34","_hasChildren":false,"_value":34}
Looking at frame number: 35
    variable 'i': {"_type":"number","_description":"35","_hasChildren":false,"_value":35}
Looking at frame number: 36
    variable 'i': {"_type":"number","_description":"36","_hasChildren":false,"_value":36}
Looking at frame number: 37
    variable 'i': {"_type":"number","_description":"37","_hasChildren":false,"_value":37}
Looking at frame number: 38
    variable 'i': {"_type":"number","_description":"38","_hasChildren":false,"_value":38}
Looking at frame number: 39
    variable 'i': {"_type":"number","_description":"39","_hasChildren":false,"_value":39}
Looking at frame number: 40
    variable 'i': {"_type":"number","_description":"40","_hasChildren":false,"_value":40}
Looking at frame number: 41
    variable 'i': {"_type":"number","_description":"41","_hasChildren":false,"_value":41}
Looking at frame number: 42
    variable 'i': {"_type":"number","_description":"42","_hasChildren":false,"_value":42}
Looking at frame number: 43
    variable 'i': {"_type":"number","_description":"43","_hasChildren":false,"_value":43}
Looking at frame number: 44
    variable 'i': {"_type":"number","_description":"44","_hasChildren":false,"_value":44}
Looking at frame number: 45
    variable 'i': {"_type":"number","_description":"45","_hasChildren":false,"_value":45}
Looking at frame number: 46
    variable 'i': {"_type":"number","_description":"46","_hasChildren":false,"_value":46}
Looking at frame number: 47
    variable 'i': {"_type":"number","_description":"47","_hasChildren":false,"_value":47}
Looking at frame number: 48
    variable 'i': {"_type":"number","_description":"48","_hasChildren":false,"_value":48}
Looking at frame number: 49
    variable 'i': {"_type":"number","_description":"49","_hasChildren":false,"_value":49}
Tests done
